Debug code and DoubleStep fix
POTENTIAL BUG:
It sometimes happens that the overall winning percentage is
higher than that of the best move, which is a bit... WTF.
The problem is probably that we reverse the score one time to much

E.g. root has 6 wins, 10 games and both children have 2 wins and
5 games played. Reverse would be exactly what root has now,
which is frightening.