Add multiplayer, multiphase, or variable turn order support.
Adds a method to the game_state interface to ask it who the winner is, rather than if the current player won. Uses this during backpropagation, increasing the value of the node if the move was made by the acting player. The main use-case for this is with multiplayer games. For instance, with a game where players are red, blue, and green, the red player wants to maximize a move while minimizing two following moves by blue and green.

With multi-phase turns, for example we might have a game of blackjack, where a player can hit, hit, hit, and stay. Even though we don't wait for the dealer to act, each hit represents a decision made by the player. This could be generalized as the player choosing between 1, 2, 3, 4, or 5 hits before being forced to stay, however between some of those might be buying insurance or doubling down. It's simpler and probably better to model the game based on how the actual game works, and the state being an intermediate between each action/move.

Since we aren't just flipping the `won` boolean, variable turn order is also supported.